requirements.txt
boto3==1.34.73
botocore==1.34.73
click==8.1.7
qdrant-client==1.8.2
python-dotenv==1.0.1
oss2==2.18.4
tabulate==0.9.0